ENH add dtype for MultiNiftiMasker

+++ nilearn/input_data/multi_nifti_masker.py
@@ -269,8 +269,7 @@ class MultiNiftiMasker(NiftiMasker, CacheMixin):
                                        target_fov=target_fov,
                                        memory=self.memory,
                                        memory_level=self.memory_level,
-                                       verbose=self.verbose,
-                                       dtype=self.dtype)
+                                       verbose=self.verbose)
 
         if confounds is None:
             confounds = itertools.repeat(None, len(imgs_list))
